001-addr-obj/requirements.txt
pan-os-python
fastmcp
setuptools